Testing on Windows 7 64-bit with MinGW-w64
Some problems arise because sizeof(long)=4 and sizeof(size_t)=8

Fixed it and everything seems to work apart from the tests that were
also failing on Windows 7 32-bit.
One bigger problem that I avoided by skipping the test on this
particular platform is fit.F90: because of fgsl_long and fgsl_size_t
confusion, it doesnt compile anymore. We may have to come up with some
generic functions or just rewrite the example.